Fix crash regression due to ABI incompatibility
Fix the regression due to the following commit:
https://github.com/notepad-plus-plus/notepad-plus-plus/commit/3e9413d6152db40e1cf31ecc8f2dd057451c998f
Modify type of "PositionCR" to "int64_t" from "long" again, to keep ABI compatible to plugins.

Ref: https://github.com/notepad-plus-plus/notepad-plus-plus/pull/12327#issuecomment-1297906305

+++ scintilla/include/Sci_Position.h
@@ -18,7 +18,7 @@ typedef ptrdiff_t Sci_Position;
 typedef size_t Sci_PositionU;
 
 // For Sci_CharacterRange  which is defined as long to be compatible with Win32 CHARRANGE
-typedef long Sci_PositionCR;
+typedef intptr_t Sci_PositionCR;
 
 #ifdef _WIN32
 	#define SCI_METHOD __stdcall
